fix(Next): Move files to library to prepare for Components
