+++ api/server/plugins/home/lib/index.js
@@ -1,4 +1,4 @@
-const gallery = require('../../../../../app/src/galleries');
+const gallery = require('../../../../../app/src/lib/galleries');
 
 async function handler(request, reply) {
   const isRaw = request.query.raw;
