+++ api/server/plugins/resize/test/index.spec.js
@@ -1,6 +1,6 @@
 const tape = require('tape-catch');
 
-const existsChecker = require('../../../../../app/src/exists');
+const existsChecker = require('../../../../../app/src/lib/exists');
 
 tape('Verify /resize route', { skip: false }, (describe) => {
   const calipers = require('calipers')('jpeg');
